Split  SETTING_ACTIVE_ACTIONS into SETTING_ACTIVE_KEY_ACTIONS and SETTING_ACTIVE_KEYS_ACTIONS Restore tests respect if store suppors setting active key/keys.
