+++ tests/state/Areas/AreaTreeLevels/actions/_actions-test.js
@@ -1,4 +1,6 @@
-import commonActionsTests from '../../../_common/actions/';
+import commonActionsTests, {
+	SETTING_ACTIVE_KEY_ACTIONS,
+} from '../../../_common/actions/';
 import actions from '../../../../../src/state/Areas/AreaTreeLevels/actions';
 import testBatchRunner, {
 	getDispatchedActionsModificator,
@@ -6,11 +8,11 @@ import testBatchRunner, {
 } from '../../../helpers';
 
 const actionNames = [
+	SETTING_ACTIVE_KEY_ACTIONS,
 	'refreshUses',
 	'useIndexed',
 	'useIndexedClear',
 	'useKeys',
-	'setActiveKey',
 ];
 
 const store = 'AREAS.AREA_TREE_LEVELS';
