+++ tests/state/Scopes/actions/_actions-test.js
@@ -1,7 +1,7 @@
 import commonActionsTests, {
 	USE_ACTIONS,
 	EDIT_ACTIONS,
-	SETTING_ACTIVE_ACTIONS,
+	SETTING_ACTIVE_KEY_ACTIONS,
 	RESTORE_STATE_ACTIONS,
 } from '../../_common/actions/';
 import actions from '../../../../src/state/Scopes/actions';
@@ -13,8 +13,8 @@ import testBatchRunner, {
 const actionNames = [
 	...USE_ACTIONS,
 	...EDIT_ACTIONS,
-	...SETTING_ACTIVE_ACTIONS,
-	...RESTORE_STATE_ACTIONS,
+	...SETTING_ACTIVE_KEY_ACTIONS,
+	...RESTORE_STATE_ACTIONS.withSetActiveKey,
 ];
 
 const store = 'SCOPES';
