+++ tests/state/Tags/actions/_actions-test.js
@@ -1,7 +1,6 @@
 import commonActionsTests, {
 	USE_ACTIONS,
 	EDIT_ACTIONS,
-	SETTING_ACTIVE_ACTIONS,
 	RESTORE_STATE_ACTIONS,
 } from '../../_common/actions/';
 import actions from '../../../../src/state/Tags/actions';
@@ -10,7 +9,11 @@ import testBatchRunner, {
 	getTestsByActionName,
 } from '../../helpers';
 
-const actionNames = [...USE_ACTIONS, ...EDIT_ACTIONS, ...RESTORE_STATE_ACTIONS];
+const actionNames = [
+	...USE_ACTIONS,
+	...EDIT_ACTIONS,
+	...RESTORE_STATE_ACTIONS.withSetActiveKey,
+];
 
 const store = 'TAGS';
 const dataType = 'tags';
